preemptive commit

+++ command/MoveFileCommand.rb
@@ -17,7 +17,7 @@ class MoveFileCommand < Command
   def execute
 
     #does the file exist and is it a file?
-    if File.exists? path and !File.directory? path
+    if File.exists? path and !File.directory? path and Dir.exists? newLoc
       
       #store the text
       file = File.open(path, "r")
